test: switch options menu tests to react testing library
